refactor(agent): extract ImplementAndJudge chain into implement_judge module
Move the implementer→judge chain (~550 lines: regexes, params,
fingerprint, verdict parsing, prompt builders, ImplementAndJudgeTool)
out of tools/agent/__init__.py into tools/agent/implement_judge.py.
The full previous import surface of pythinker_code.tools.agent is
preserved via explicit re-exports, so the agent-spec tool path and all
existing imports work unchanged; the new submodule is registered in the
PyInstaller hiddenimports snapshot.
